+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.14.5-rc.1",
-  genVersion: "2.731.9",
-  userAgent: "speakeasy-sdk/typescript 3.14.5-rc.1 2.731.9 2.0 @orq-ai/node",
+  sdkVersion: "4.0.0-rc.0",
+  genVersion: "2.735.1",
+  userAgent: "speakeasy-sdk/typescript 4.0.0-rc.0 2.735.1 2.0 @orq-ai/node",
 } as const;